+++ examples/next-hello-world/package.json
@@ -14,6 +14,7 @@
     "@trpc/client": "^6.0.0-alpha.5",
     "@trpc/react": "^6.0.0-alpha.16",
     "@trpc/server": "^6.0.0-alpha.5",
+    "@trpc/next": "^6.0.0-alpha.5",
     "next": "^10.0.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
