+++ examples/next-hello-world/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { withTRPCClient } from '@trpc/react';
+import { withTRPCClient } from '@trpc/next';
 import { AppType } from 'next/dist/next-server/lib/utils';
 import React from 'react';
 import type { AppRouter } from './api/trpc/[trpc]';
@@ -9,7 +9,7 @@ const MyApp: AppType = ({ Component, pageProps }) => {
 
 MyApp.getInitialProps = async ({ ctx }) => {
   // make it sloow
-  await new Promise((resolve) => setTimeout(resolve, 1000));
+  // await new Promise((resolve) => setTimeout(resolve, 1000));
 
   const ONE_DAY_SECONDS = 60 * 60 * 24;
   ctx.res?.setHeader(
@@ -20,24 +20,26 @@ MyApp.getInitialProps = async ({ ctx }) => {
     pageProps: {},
   };
 };
-const baseUrl = process.browser
-  ? ''
-  : process.env.VERCEL_URL
-  ? `https://${process.env.VERCEL_URL}`
-  : 'http://localhost:3000';
 
-const url = `${baseUrl}/api/trpc`;
-
-console.log('url', url);
-export default withTRPCClient<AppRouter>(MyApp, {
-  url,
-  ssr: true,
-  getHeaders() {
+export default withTRPCClient<AppRouter>(
+  ({ ctx }) => {
     if (process.browser) {
-      return {};
+      return {
+        url: '/api/trpc',
+      };
     }
+    const url = process.env.VERCEL_URL
+      ? `https://${process.env.VERCEL_URL}/api/trpc`
+      : 'http://localhost:3000/api/trpc';
+
     return {
-      'x-ssr': '1',
+      url,
+      getHeaders() {
+        return {
+          'x-ssr': '1',
+        };
+      },
     };
   },
-});
+  { ssr: true },
+)(MyApp);
