--- packages/client/CHANGELOG.md
@@ -1,1 +1,0 @@
-# `@trpc/client`
