+++ packages/next/README.md
@@ -1,0 +1,3 @@
+# `@trpc/react`
+
+https://trpc.io
