+++ packages/next/test/README.md
@@ -1,0 +1,2 @@
+See [`../../server/test`](../../server/test) - we test this lib there too
+
