--- packages/react/CHANGELOG.md
@@ -1,1 +1,0 @@
-# @thefakeorg/react
