+++ packages/react/src/index.ts
@@ -1,3 +1,2 @@
 export * from '@trpc/client';
 export * from './createReactQueryHooks';
-export * from './withTRPCClient';
