--- packages/server/CHANGELOG.md
@@ -1,2 +1,0 @@
-# `@trpc/server`
-
